+++ packages/core/src/vuePlugins/actions/useExtendedMasterportalapiMarkers/index.ts
@@ -2,7 +2,7 @@ import { Feature, MapBrowserEvent } from 'ol'
 import {
   CoreGetters,
   CoreState,
-  ExtendedMasterportalapiMarkersIsSelectableFunction,
+  ExtendedMasterportalapiMarkers,
   MarkerStyle,
   PolarActionContext,
   PolarStore,
@@ -80,15 +80,7 @@ export function useExtendedMasterportalapiMarkers(
     layers,
     clusterClickZoom = false,
     dispatchOnMapSelect,
-  }: {
-    hoverStyle?: MarkerStyle
-    selectionStyle?: MarkerStyle
-    unselectableStyle?: MarkerStyle
-    isSelectable?: ExtendedMasterportalapiMarkersIsSelectableFunction
-    layers: string[]
-    clusterClickZoom: boolean
-    dispatchOnMapSelect?: string[]
-  }
+  }: ExtendedMasterportalapiMarkers
 ) {
   localSelectionStyle = selectionStyle
   const { map } = getters
@@ -102,7 +94,7 @@ export function useExtendedMasterportalapiMarkers(
     .filter(layerFilter)
     .forEach((layer) => {
       // only vector layers reach this
-      const source = (layer as VectorLayer<Feature>).getSource()
+      const source = (layer as VectorLayer).getSource()
       if (source !== null) {
         // @ts-expect-error | Undocumented hook.
         source.geometryFunction =
@@ -110,8 +102,8 @@ export function useExtendedMasterportalapiMarkers(
           (feature: Feature) =>
             isVisible(feature) ? feature.getGeometry() : null
       }
-      const originalStyleFunction = (layer as VectorLayer<Feature>).getStyle()
-      ;(layer as VectorLayer<Feature>).setStyle((feature) => {
+      const originalStyleFunction = (layer as VectorLayer).getStyle()
+      ;(layer as VectorLayer).setStyle((feature) => {
         if (
           typeof isSelectable === 'undefined' ||
           isSelectable(feature as Feature)
@@ -207,7 +199,6 @@ export function useExtendedMasterportalapiMarkers(
       setLayerId(map, feature)
       selected = feature
       if (dispatchOnMapSelect) {
-        // @ts-expect-error | May be one or two elements, no fixed tuple.
         dispatch(...dispatchOnMapSelect)
       }
       hovered?.setStyle?.(undefined)