+++ packages/lib/passesBoundaryCheck/index.ts
@@ -1,4 +1,4 @@
-import { Feature, Map } from 'ol'
+import { Map } from 'ol'
 import { Coordinate } from 'ol/coordinate'
 import VectorLayer from 'ol/layer/Vector'
 import VectorSource from 'ol/source/Vector'
@@ -50,9 +50,7 @@ export const passesBoundaryCheck = async (
   const boundaryLayer = map
     .getLayers()
     .getArray()
-    .find(
-      (layer) => layer.get('id') === boundaryLayerId
-    ) as VectorLayer<Feature>
+    .find((layer) => layer.get('id') === boundaryLayerId) as VectorLayer
 
   if (!boundaryLayer) {
     console.error(