+++ packages/plugins/Draw/src/store/createInteractions/createDeleteInteractions.ts
@@ -3,11 +3,10 @@ import VectorLayer from 'ol/layer/Vector'
 import Interaction from 'ol/interaction/Interaction'
 import { DragBox, Select } from 'ol/interaction'
 import { platformModifierKeyOnly } from 'ol/events/condition'
-import { Feature } from 'ol'
 
 export default function (
   drawSource: VectorSource,
-  drawLayer: VectorLayer<Feature>
+  drawLayer: VectorLayer
 ): Interaction[] {
   const selectInteraction = new Select({ layers: [drawLayer] })
   const selectedFeatures = selectInteraction.getFeatures()
@@ -29,7 +28,7 @@ export default function (
     selectedFeatures.clear()
   })
 
-  // @ts-expect-error | internal hack to detect it in Pins plugin
+  // @ts-expect-error | internal hack to detect it in other plugins
   selectInteraction._isDeleteSelect = true
 
   return [selectInteraction, dragBoxInteraction]