+++ packages/lib/getFeatures/CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## unpublished
+
+- Breaking: Upgrade peerDependency `ol` from `^9.2.4` to `^10.3.1`.
+
 ## 2.0.0
 
 - Breaking: Upgrade peerDependency of `ol` from `^7.1.0` to `^9.2.4`.