+++ packages/plugins/Pins/CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## unpublished
+
+- Breaking: Upgrade peerDependency `ol` from `^9.2.4` to `^10.3.1`.
+- Feature: Recognize the multi-select-`Draw` of `@polar/plugin-gfi`.
+
 ## 2.0.0
 
 - Breaking: Upgrade peerDependency `ol` from `^7.1.0` to `^9.2.4`.